Small improvements, cleanup

+++ create_sys_image.sh
@@ -1,2 +1,4 @@
-cp MakieSys.so MakieSys.so.bak
+if test -f "MakieSys.so"; then
+    cp MakieSys.so MakieSys.so.bak
+fi
 julia --project  -e "include(\"./test/create_sys_image.jl\");"
